Moved sanitizing to external class from Driver
